clean up and add distributed file read
