--- test/library/draft/arrow/ArrowDemo.skipif
@@ -1,4 +1,0 @@
-#!/usr/bin/env python3
-
-import os
-print(os.getenv('ARROW_IMPL') == 'None')
