+++ test/library/draft/arrow/lib/null.c
@@ -1,0 +1,9 @@
+#include<null.h>
+
+int isNull(void *ptr){
+    return ptr == NULL ? 1 :0;
+}
+
+void* getNULL(){
+    return NULL;
+}
