+++ test/library/draft/arrow/lib/null.h
@@ -1,0 +1,11 @@
+#ifndef ARRAY_H
+#define ARRAY_H
+
+#include<arrow-glib/arrow-glib.h>
+
+
+int isNull(void*);
+
+void* getNULL(void);
+
+#endif // ARRAY_H
