--- test/library/draft/arrow/null.c
@@ -1,9 +1,0 @@
-#include<null.h>
-
-int isNull(void *ptr){
-    return ptr == NULL ? 1 :0;
-}
-
-void* getNULL(){
-    return NULL;
-}
