+++ test/library/draft/arrow/test/CreateParquet.good
@@ -1,0 +1,2 @@
+0 1 2 3 4 5 6 7 8 9
+0 1 2 3 4 5 6 7 8 9
